# dashboard/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Replication Manager Monitoring and CLI for MariaDB and MySQL",
    "title": "replication-manager",
    "contact": {
      "name": "Stephane Varoqui",
      "email": "svaroqui@gmail.com"
    },
    "license": {
      "name": "GPL",
      "url": "http://opensource.org/licenses/GPL"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/",
  "paths": {
...
    "/api/clusters/{clusterName}/status": {
      "get": {
        "description": "Shows the status for that specific named cluster",
        "operationId": "clusterStatus",
        "parameters": [
          {
            "type": "string",
            "description": "cluster to filter by",
            "name": "clusterName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/status"
          }
        }
      }
    },
...
  },
...
  "responses": {
...
    "status": {
      "description": "The Status contains string value for the alive status.\nPossible values are: running, starting, errors",
      "schema": {
        "type": "object",
        "properties": {
          "alive": {
            "type": "string",
            "x-go-name": "Alive",
            "example": "running"
          }
        }
      },
      "headers": {
        "Access-Control-Allow-Origin": {
          "type": "string",
          "example": "*"
        }
      }
    },
...
  },
  "security": [
    {
      "api_key": []
    }
  ]
}